pyproject.toml
[tool.poetry]
name = "hw-24-1"
version = "0.1.0"
description = ""
authors = ["UltimaKIND <minion9@mail.ru>"]
readme = "README.md"
packages = [{include = "hw_24"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.1"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
black = "^24.8.0"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
